removed another pointless abstraction
