--- src/Ocelot/Configuration/Authentication/IHashMatcher.cs
@@ -1,7 +1,0 @@
-namespace Ocelot.Configuration.Authentication
-{
-    public interface IHashMatcher
-    {
-        bool Match(string password, string salt, string hash);
-    }
-}
